% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpdns.R
\name{pdns_query}
\alias{pdns_query}
\alias{pdns.query}
\title{Query CIRCL.LU Passive DNS API}
\usage{
pdns_query(domain, circl_user = Sys.getenv("CIRCL_USER"),
  circl_password = Sys.getenv("CIRCL_PASSWORD"), use_proxy,
  circl_pdns_service_url = "https://www.circl.lu/pdns/query/",
  enable_cache = FALSE)

pdns.query(domain, circl_user = Sys.getenv("CIRCL_USER"),
  circl_password = Sys.getenv("CIRCL_PASSWORD"), use_proxy,
  circl_pdns_service_url = "https://www.circl.lu/pdns/query/",
  enable_cache = FALSE)
}
\arguments{
\item{domain}{host/domain name to query}

\item{circl_user, circl_password}{circl.lu service username/password. Highly suggested
that these be stored in environment variables (see parameters for which ones).}

\item{use_proxy}{switch proxy on, has to be an \code{httr::use_proxy} object, e.g.
\code{use_proxy(url, port, username = "", password = "")}}

\item{circl_pdns_service_url}{<circl.lu> passive DNS service URL}

\item{enable_cache}{switch API result caching on / off}
}
\description{
Given a domain name this will return a data frame consisting of passive DNS entries.
}
\section{TODO}{

\itemize{
\item Expiring cache
\item Domain age
}
}

\references{
CIRCL.LU's \href{https://github.com/CIRCL/PyPDNS}{PyPDNS Python module}
}
