% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpdns-package.R
\docType{package}
\name{rpdns}
\alias{rpdns}
\alias{rpdns-package}
\title{Query 'CIRCL' Passive 'DNS' 'API'}
\description{
'CIRCL' 'Passive DNS' \url{https://www.circl.lu/services/passive-dns/} is a
database storing historical 'DNS' records from various resources including malware
analysis or partners. The 'DNS' historical data is indexed, which makes it searchable
for incident handlers, security analysts or researchers.
}
\author{
Bob Rudis (bob@rud.is)
}
